Release v1.0.4: Change amount fields from decimal to long (smallest currency unit)

+++ src/Sunbay.Nexus.Sdk/Models/Common/AuthAmount.cs
@@ -6,14 +6,15 @@ namespace Sunbay.Nexus.Sdk.Models.Common
     /// Authorization amount information
     /// Supports: orderAmount, pricingCurrency only
     /// Used for: Auth, ForcedAuth, IncrementalAuth
+    /// All amounts are in the smallest currency unit (e.g., cents for USD, fen for CNY)
     /// </summary>
     public class AuthAmount
     {
         /// <summary>
-        /// Order amount (required)
+        /// Order amount in smallest currency unit (required)
         /// </summary>
         [JsonPropertyName("orderAmount")]
-        public decimal OrderAmount { get; set; }
+        public long OrderAmount { get; set; }
         
         /// <summary>
         /// Pricing currency (ISO 4217, required)
